+++ src/bump_scope_guard.rs
@@ -1,8 +1,10 @@
 use core::{fmt::Debug, marker::PhantomData, num::NonZeroUsize, ptr::NonNull};
 
 use crate::{
-    chunk_header::ChunkHeader, polyfill::nonnull, BaseAllocator, Bump, BumpScope, MinimumAlignment, RawChunk, Stats,
-    SupportedMinimumAlignment,
+    chunk_header::ChunkHeader,
+    polyfill::nonnull,
+    stats::{AnyStats, Stats},
+    BaseAllocator, Bump, BumpScope, MinimumAlignment, RawChunk, SupportedMinimumAlignment,
 };
 
 /// This is returned from [`checkpoint`](Bump::checkpoint) and used for [`reset_to`](Bump::reset_to).
@@ -42,7 +44,7 @@ where
     A: BaseAllocator,
 {
     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        self.stats().debug_format("BumpScopeGuard", f)
+        AnyStats::from(self.stats()).debug_format("BumpScopeGuard", f)
     }
 }
 
@@ -94,7 +96,7 @@ where
     /// Returns a type which provides statistics about the memory usage of the bump allocator.
     #[must_use]
     #[inline(always)]
-    pub fn stats(&self) -> Stats<'a, true> {
+    pub fn stats(&self) -> Stats<'a, A, UP> {
         let header = self.chunk.header_ptr().cast();
         // SAFETY: `header` points to a valid chunk header which is guaranteed allocated
         unsafe { Stats::from_header_unchecked(header) }
@@ -130,7 +132,7 @@ where
     A: BaseAllocator,
 {
     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        self.stats().debug_format("BumpScopeGuardRoot", f)
+        AnyStats::from(self.stats()).debug_format("BumpScopeGuardRoot", f)
     }
 }
 
@@ -179,7 +181,7 @@ where
     /// Returns a type which provides statistics about the memory usage of the bump allocator.
     #[must_use]
     #[inline(always)]
-    pub fn stats(&self) -> Stats {
+    pub fn stats(&self) -> Stats<A, UP> {
         let header = self.chunk.header_ptr().cast();
         // SAFETY: `header` points to a valid chunk header which is guaranteed allocated
         unsafe { Stats::from_header_unchecked(header) }