+++ src/lib.rs
@@ -333,7 +333,6 @@ mod raw_fixed_bump_string;
 mod raw_fixed_bump_vec;
 mod set_len_on_drop;
 mod set_len_on_drop_by_ptr;
-/// Types that provide statistics about the memory usage of the bump allocator.
 pub mod stats;
 mod without_dealloc;
 
@@ -371,8 +370,6 @@ pub use no_drop::NoDrop;
 use private::{capacity_overflow, format_trait_error, PanicsOnAlloc};
 use raw_chunk::RawChunk;
 use set_len_on_drop::SetLenOnDrop;
-#[doc(inline)]
-pub use stats::Stats;
 pub use without_dealloc::{WithoutDealloc, WithoutShrink};
 
 #[cfg(feature = "zerocopy-08")]
@@ -589,7 +586,7 @@ macro_rules! bump_common_methods {
                 /// Returns a type which provides statistics about the memory usage of the bump allocator.
                 #[must_use]
                 #[inline(always)]
-                pub fn stats(&self) -> Stats<'a, GUARANTEED_ALLOCATED> {
+                pub fn stats(&self) -> Stats<'a, A, UP, GUARANTEED_ALLOCATED> {
                     let header = self.chunk.get().header_ptr().cast();
                     unsafe { Stats::from_header_unchecked(header) }
                 }
@@ -597,7 +594,7 @@ macro_rules! bump_common_methods {
                 /// Returns a type which provides statistics about the memory usage of the bump allocator.
                 #[must_use]
                 #[inline(always)]
-                pub fn stats(&self) -> Stats<GUARANTEED_ALLOCATED> {
+                pub fn stats(&self) -> Stats<A, UP, GUARANTEED_ALLOCATED> {
                     let header = self.chunk.get().header_ptr().cast();
                     unsafe { Stats::from_header_unchecked(header) }
                 }
@@ -697,7 +694,7 @@ macro_rules! collection_method_allocator_stats {
         /// This merely exists for api parity with `Mut*` collections which can't have a `allocator` method.
         #[must_use]
         #[inline(always)]
-        pub fn allocator_stats(&self) -> Stats {
+        pub fn allocator_stats(&self) -> $crate::stats::AnyStats {
             self.allocator.stats()
         }
     };
@@ -712,7 +709,7 @@ macro_rules! mut_collection_method_allocator_stats {
         /// This collection does not update the bump pointer, so it also doesn't contribute to the `remaining` and `allocated` stats.
         #[must_use]
         #[inline(always)]
-        pub fn allocator_stats(&self) -> Stats {
+        pub fn allocator_stats(&self) -> $crate::stats::AnyStats {
             self.allocator.stats()
         }
     };